Graphics.UI.Widget.Core

Description

Core of Widget

Widget functions

override :: (Widget old -> Widget new) -> Widget old -> (forall br m. Union new br m ~> (br (Widget new) m `Sum` Union old br m)) -> Widget new

Recursive extension

runSwitch :: Widget xs -> Getter (Widget xs) (FreezeT (Widget xs) Identity a) -> (Freeze (Widget xs) a -> r) -> r

Combinator for running a widget with switching method like op'switch

runSwitchM :: (k ∈ xs, Monad m) => Widget xs -> k FreezeT m a -> (Freeze (Widget xs) a -> m r) -> m r

runSwitch with monad

op'isFreeze :: Widget xs -> Getter (Widget xs) (FreezeT (Widget xs) Identity a) -> Bool

Is a widget frozen?

Config

type WConfig k = WConfigR (Record (Required k)) (Record (Optional k))

WConfig k is a record of widget id, required parameter and optional parameter

class Conf k where

k has default value of optional parameters

Minimal complete definition

def

Associated Types

type Required k :: [Assoc Symbol *]

type Optional k :: [Assoc Symbol *]

Methods

def :: Record (Optional k)

default values

data ViewWConfig a b

Unpack a value of WConfig

Constructors

ViewWConfig WidgetId a b

giveWid :: KnownSymbol k => Proxy k -> WConfig k -> WConfig k

Give a widget-id using Proxy

getLocation :: Given StyleSheet => WidgetId -> V2 Int

Find the location of the widget corresponding to given widget-id

mkCfg :: WidgetId -> a -> b -> WConfigR a b

Build Config data

viewWConfig :: WConfigR a b -> ViewWConfig a b

Unpacking function

wconf :: KnownSymbol k => Proxy k -> WConfig k -> ViewWConfig (Record (Required k)) (Record (Optional k))

wconf p = viewWConfig . giveWid p

conf :: WidgetId -> Record (Required k) -> Record (Optional k) -> WConfig k

combinator for a widget with wix (widget-id)

  conf wix req opt
  = #wix = wix
  <: #required = req
  <: #optional @= opt
  <: emptyRecord

Common Operators

op'render :: (Given StyleSheet, KnownName xs, Op'Render ∈ xs) => Getter (Widget xs) (RenderM ())

op'render = op'renderAlpha 1.0

op'renderAlpha :: (KnownName xs, Op'Render ∈ xs) => Double -> Getter (Widget xs) (RenderM ())

op'run :: (∈) Op'Run xs => Getter (Widget xs) (RenderM (Widget xs))

op'reset :: Op'Reset arg ∈ xs => arg -> Getter (Widget xs) (Widget xs)

op'reset = _self' . Op'Reset

op'handleEvent :: (∈) Op'HandleEvent xs => Map Scancode Int -> Getter (Widget xs) (RenderM (Widget xs))

op'switch :: (∈) Op'Switch xs => Getter (Widget xs) (FreezeT (Widget xs) Identity ())

Re-exports

type family (k ++ (xs :: [k])) (ys :: [k]) :: [k] where ... infixr 5

Equations

(k ++ ([] k)) ys = ys

(k ++ ((:) k x xs)) ys = (:) k x ((++) k xs ys)

type (∈) k x xs = Member k xs x

Unicode flipped alias for Member

module Graphics.UI.Widget.Internal.Widget

module Graphics.UI.Widget.Internal.TH

module Graphics.UI.Widget.Internal.Named

module Graphics.UI.Widget.Internal.Freeze