Graphics.UI.Widget.Util

Description

Utility module

Lenses

functorial :: Functor f => Getter a b -> Getter (f a) (f b)

functorial lifting

monadic :: Monad m => Lens' a b -> Lens' (m a) (m b)

monadic lifting

(^%~) :: Lens' s a -> Getter a a -> s -> s infixr 4

s ^%~ f = s %~ (^. f)

(^%%~) :: Functor m => Lens' s a -> Getter a (m a) -> s -> m s infixr 4

s ^%%~ f = s %%~ (^. f)

(^%=) :: MonadState s m => Lens' s a -> Getter a a -> m () infixr 4

s ^%= f = s %= (^. f)

(^%%=) :: MonadState s m => Lens' s a -> Getter a (m a) -> m () infixr 4

s ^%%= f = use s >>= \s' -> s <~ s' ^. f

Extensible

hmerge :: (xs ⊆ ys, Wrapper h) => (h :* xs) -> (h :* ys) -> h :* ys

Merges two products when first one is included in second one

hmergeAssoc :: (IncludeAssoc ys xs, Wrapper h) => (h :* xs) -> (h :* ys) -> h :* ys

Merges two associated products when first one is included in second one